+++ common/i18n/src/test/java/is/codion/common/i18n/MessagesTest.java
@@ -41,7 +41,6 @@ public class MessagesTest {
 		Messages.clear();
 		Messages.clearTip();
 		Messages.clearMnemonic();
-		Messages.advanced();
 		Messages.find();
 		Messages.refresh();
 		Messages.refreshMnemonic();