+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/tutorial/MultiPanelDemo.java
@@ -24,6 +24,7 @@ import is.codion.framework.db.local.LocalEntityConnectionProvider;
 import is.codion.framework.demos.chinook.domain.Chinook.Employee;
 import is.codion.framework.demos.chinook.domain.impl.ChinookImpl;
 import is.codion.framework.demos.chinook.ui.EmployeeEditPanel;
+import is.codion.swing.common.ui.component.table.ColumnConditionPanel.ConditionState;
 import is.codion.swing.common.ui.dialog.Dialogs;
 import is.codion.swing.framework.model.SwingEntityModel;
 import is.codion.swing.framework.ui.EntityPanel;
@@ -55,7 +56,7 @@ public final class MultiPanelDemo {
 		JPanel basePanel = new JPanel(gridLayout(2, 2));
 		for (int i = 0; i < 4; i++) {
 			EntityPanel employeePanel = new EntityPanel(employeeModel, new EmployeeEditPanel(employeeModel.editModel()));
-			employeePanel.tablePanel().conditionPanelVisible().set(true);
+			employeePanel.tablePanel().conditionPanel().state().set(ConditionState.SIMPLE);
 			employeePanel.initialize();
 			basePanel.add(employeePanel);
 		}